requirements.txt
aiohttp>=3.8.3
PyYAML>=6.0